more comments and restore script

+++ install.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-
+set -x
 fatal(){
   echo "$@"
   exit 1
@@ -19,12 +19,12 @@ ln -sf /config/letsencrypt/letsrenew.sh /etc/cron.monthly/letsrenew.sh
 curl -o /config/scripts/post-config.d/install_letsencrypt.sh https://raw.githubusercontent.com/rholmboe/ubnt-letsencrypt/master/install_letsencrypt.sh
 chmod 755 /config/scripts/post-config.d/install_letsencrypt.sh
 
-# Generate certifications which will be used
-openssl genrsa 4096 | tee /config/letsencrypt/account.key
-openssl genrsa 4096 | tee /config/letsencrypt/domain.key
-openssl req -new -sha256 -key /config/letsencrypt/domain.key -subj "/CN=$fqdn" | tee /config/letsencrypt/domain.csr
+echo "Generate keys to be used in our signed certificate"
+[ -f /config/letsencrypt/account.key ] || openssl genrsa 4096 | tee /config/letsencrypt/account.key
+[ -f /config/letsencrypt/domain.key ]  || openssl genrsa 4096 | tee /config/letsencrypt/domain.key
+[ -f /config/letsencrypt/domain.csr ]  || openssl req -new -sha256 -key /config/letsencrypt/domain.key -subj "/CN=$fqdn" | tee /config/letsencrypt/domain.csr
 
-# Making lighttpd configurations and restarting daemon
+echo "Making lighttpd configurations and restarting daemon so letsencypt and verify we own this domain via .well-known/acme-challenge/"
 [ -d /config/lighttpd ] ||  mkdir /config/lighttpd/
 curl -o /config/lighttpd/lighttpd.conf https://raw.githubusercontent.com/rholmboe/ubnt-letsencrypt/master/lighttpd.conf
 ln -sf /config/lighttpd/lighttpd.conf /etc/lighttpd/lighttpd.conf
@@ -34,7 +34,10 @@ ps -e | grep lighttpd | awk '{print $1;}' | xargs kill
 # Create acme respons directory
 [ -d /var/www/htdocs/.well-known/acme-challenge/ ] || mkdir -p /var/www/htdocs/.well-known/acme-challenge/
 
-# Run letsrenew.sh file for initial connect and/or renewal, doesn't matter
+echo "Run letsrenew.sh file to actually sign the certificate and generate new /etc/lighttpd/server.pem"
 bash /config/letsencrypt/letsrenew.sh
+
+
+
